requirements.txt
# ── Streamlit ─────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── LangChain stack ───────────────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.5
langchain-huggingface>=0.0.3

# ── PDF loading ───────────────────────────────────────────────────────────────
pypdf>=4.0.0

# ── Embeddings (local, CPU-safe) ──────────────────────────────────────────────
sentence-transformers>=2.7.0
torch>=2.0.0

# ── Vector store ─────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── Google Gemini ─────────────────────────────────────────────────────────────
google-generativeai>=0.7.0
